add release for locks

+++ src/main/java/com/redis/runner/Reader.java
@@ -4,6 +4,7 @@ import com.redis.Launcher;
 import com.redis.repository.model.Message;
 import com.redis.service.UpdateService;
 import java.util.List;
+import java.util.Random;
 
 public class Reader implements Runnable {
 
@@ -21,11 +22,17 @@ public class Reader implements Runnable {
                 .ifPresent(id -> {
                     List<Message> updates = updateService.readAllUpdates(group, id);
                     updates.forEach(System.out::println);
+                    if (new Random().nextInt(100) > 92) {
+                        System.out.println("simulated error for group = " + group + " with id = " + id);
+                        return;
+                    }
                     boolean handled = updateService.deleteOldUpdates(group, id, updates.size());
 
                     if (!handled) {
                         System.out.println("group " + group + " with id " + id + " size = " + updates.size());
                         System.out.println("group " + group + " with id " + id + " " + " was not handled");
+                    } else {
+                        System.out.println("group " + group + " with id " + id + " " + " handled success");
                     }
                 });
